requirements.txt
PyQt5
pyinstaller
pytest
pdfplumber
pymupdf
openpyxl
python-docx
striprtf
odfpy
odfdo
mrkdwn-analysis
Pillow
nltk==3.9.1
numpy
psd-tools
PyDriller
fpdf2
tabulate
pandas
tree-sitter-python
tree-sitter-javascript
tree-sitter-c
tree-sitter-cpp
tree-sitter-java
tree-sitter-typescript
tree-sitter-go
tree-sitter-rust